+++ app/routes/decks/$id/practice.tsx
@@ -18,10 +18,12 @@ import {
 import { cls } from "../../../utils/misc";
 import type { PageHandle } from "../../../utils/page-handle";
 import { BookmarkEntryComponent } from "../../bookmarks";
-
 import type { LoaderData } from "./_utils.server";
-import { QueueStatisticsComponent } from "./index";
-import { DeckNavBarMenuComponent } from "./index.utils";
+import {
+  DeckNavBarMenuComponent,
+  QueueStatisticsComponent,
+} from "./index.utils";
+
 export { loader } from "./_utils.server";
 
 export const handle: PageHandle = {
